+++ packages/dara-core/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 ## NEXT
 
 -  JS: Fixed an issue where passing a nullish component to `DynamicComponent` would cause the component to error out, it will now render `null` instead.
+-  Added a new `run_task` method to `ActionCtx` to imperatively run a task in a separate process.
 
 ## 1.16.20